Refactored pig-hunt 4.0 to consider parallel threads in hacking speed calculation.
We moved the number of active threads running into to the calculation functions for how many threads are currently needed to target a server effectively.

We are now calculating grow and weaken time based on parallel threads, so if we have enough threads to totally grow or weaken a server, it counts as only a single timing call because all the threads are running in parallel.