gang-manager - Fixed bug with wanted level penalty getting stuck when evaluating new gang
We discovered an issue where a new gang will always have a wanted level of 1 and a wanted penalty of 50%. This does not change until a gang member commits a crime.

The result is that the gang will get stuck in Wanted Level Reduction mode and perform Ethical Hacking forever.

We now check if the gang meets the criteria for a new gang and force the gang to prioritize crime until the initial wanted pentalty updates.