Added Upgrade Locking for Personal Servers
We were running into an issue where servers that were in the middle of a long duration hack would be upgraded and have their progress nuked. We have introduced the concept of locking a server so that new threads cannot be run on it.

When we identify that a personal server is ready for upgrade, and we can afford that upgrade, we lock the server and wait for it's current set of allocated threads to end. Once it is no longer running threads, it is upgraded and unlocked. This allows us to upgrade servers without losing hack progress.