pig-hunt-4.0.js - added support for server farm ram preallocation based on allocation access keys
Added a system that allows us to preallocate server farm ram ahead of time. Each preallocation is associated with an allocation key, if the key is not provided when DistributeScriptsToNetwork is called, that ram will not allocated. This essentially allows us to pre partion ram for tasks.

Currently, we are using this to preallocate server farm space for servers that have already been compromised by the pig hunt algorithm.

Currently, we don't support ram prallocation on the home pc, but we could use it in the future to partition out tasks, or even flag pservers as special servers for running specific opperations.